Expanded Ext, further tweaks and fixes
The system in place was not working with current definitions of Ext.
As such, I added a special ExtFix form, denoted by X bit being set.
This provides much needed fixing for dynamic instructions, now that
Push is capable of supporting them well.

Also made a few small changes to documentation and execution
that are not significant.